requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-environ==0.11.2
django-cors-headers==4.3.1

# Production server
gunicorn==21.2.0
 
# Serves static files in production without a CDN
whitenoise==6.6.0